modifed some changes to correct data issues
